phase_definitions table nonexistent, implementation removed
This one did appear to be in use, but seeing as the table does not
exist, I have removed its implementation within the server core.

There are a number of tables that exist within Zero, One and Two, that
do not exist in Four, so I will start implementing those. They may be
related to this phase_definitions usage.